clarify inline comment in Duplicate Layers.lua
It seems it's not possible to have this script operate on multiple
layers.

app.range.layers will give us all the layers that are selected, but only
one of those will be the activeLayer.

app.command.DuplicateLayer() only duplicates the activeLayer.

There's not a way to change the activeLayer using the Lua API.